Catch errors saving patient and visit
Previously only ValidationErrors were returned to the user.

Some errors like removing part of a date lead to both a validation error and a failure to save. This change means the latter doesn't bsod the page and the user can see (a very unfriendly) description of what is wrong.

This also means we don't need to bother trying to guess whether an error would cause the Patient or Visit to fail to save.