Only use our postcodes.io and don't accidentall hide unexpected exceptions.
Actually pass the postcodes.io API key! I didn't notice this stopped working because I was testing with the API key disabled in Azure APIM.

As we don't use a TaskGroup we must re-raise exceptions that occur within our async tasks. Because we already checked for ValidationError, this had the unfortunate effect of silently eating any errors that weren't HTTPError.

I noticed this because I tried to access `response.status` which doesn't exist (it's actually `response.status_code`). But I didn't see anything in the log.