Fix running one test missing mandatory field
This was just an oopsie - I was using the entire dummy_sheet.csv so obviously the assertion was going to fail! Only the first invalid row would fail to save and the other valid ones would.

The subsequent tests (missing dob etc) fail because the PDU is missing. But honestly not sure how it's there in the first place?