Fix running multiple tests missing mandatory field
The first one worked because we ran the seed_users fixture which spawns the appropriate PDU for the user if it doesn't exist already.
The problem is that was a session fixture, so it didn't run it again on the next test run.
No biggie for normal tests as the database is rolled back to the state after the fixture ran, including the PDU.
But for the transactional tests that doesn't work as it truncates all the tables after each test.

So my solution is to run the fixtures before each test (function fixtures). It's slow but still only a 40 second test run so not a big deal and I think it will be useful to have all the missing mandatory fields covered.